More JSON API tweaks & add multiCooldown to abstract windDash and windJump
