+++ src/main/kotlin/io/github/thebroccolibob/soulorigins/datagen/builder/JsonBuilder.kt
@@ -5,11 +5,11 @@ import com.google.gson.JsonElement
 import com.google.gson.JsonNull
 import com.google.gson.JsonObject
 
-class JsonObjectBuilder(val jsonObject: JsonObject) {
-    @Deprecated("Second value is not a JSON element.")
-    infix fun String.to (other: Any) {
+typealias JsonInit = JsonObjectBuilder.() -> Unit
 
-    }
+class JsonObjectBuilder(val jsonObject: JsonObject = JsonObject()) {
+    @Deprecated("Warning: Second value is not a JSON element.")
+    infix fun String.to (other: Any) {}
 
     infix fun String.to(value: String) {
         jsonObject.addProperty(this, value)
@@ -35,7 +35,7 @@ class JsonObjectBuilder(val jsonObject: JsonObject) {
         jsonObject.add(this, JsonNull.INSTANCE)
     }
 
-    inline infix fun String.to(valueInit: JsonObjectBuilder.() -> Unit) {
+    inline infix fun String.to(valueInit: JsonInit) {
         to(JsonObject(valueInit))
     }
 
@@ -101,7 +101,7 @@ class JsonObjectBuilder(val jsonObject: JsonObject) {
 }
 
 inline fun JsonObject(init: JsonObjectBuilder.() -> Unit): JsonObject {
-    return JsonObjectBuilder(JsonObject()).apply(init).jsonObject
+    return JsonObjectBuilder().apply(init).jsonObject
 }
 
 fun JsonArray(vararg jsonInit: JsonObjectBuilder.() -> Unit) = jsonInit.map(::JsonObject).toJson()
@@ -110,8 +110,6 @@ fun JsonArray(vararg elements: JsonElement) = elements.asIterable().toJson()
 fun JsonArray(vararg elements: String) = elements.asIterable().toJson()
 fun JsonArray(vararg elements: Number) = elements.asIterable().toJson()
 fun JsonArray(vararg elements: Boolean) = elements.asIterable().toJson()
-
-operator fun JsonArray.plus(other: JsonArray) = plus(other as Iterable<JsonElement>).toJson()
 
 fun Iterable<JsonElement>.toJson() = JsonArray().apply {
     this@toJson.forEach(::add)
@@ -131,3 +129,5 @@ fun Iterable<Number>.toJson() = JsonArray().apply {
 fun Iterable<Boolean>.toJson() = JsonArray().apply {
     this@toJson.forEach(::add)
 }
+
+fun listOfJson(vararg jsonInit: JsonInit) = jsonInit.map(::JsonObject)
